requirements.txt
# AURORA v4.0 — Python Dependencies (Autonomous Personality)
# Install with: pip install -r requirements.txt

# Core Audio / Voice
SpeechRecognition>=3.10.0
pyttsx3>=2.90
edge-tts
pygame>=2.5.0

# Vision & Hologram
opencv-python>=4.8.0
numpy>=1.24.0

# Web & Networking
requests>=2.31.0
websockets>=12.0
python-dotenv>=1.0.0

# AI / PyTorch (for hologram_processor.py MiDaS)
timm
torch
torchvision

# Windows OS Integration
pywin32; platform_system=="Windows"

# Optional Cloud LLM Backends
# groq>=0.4.1              # Uncomment for Groq cloud API
# huggingface-hub>=0.17.0  # Uncomment for HuggingFace

# NOTE: Local AI — Install separately:
#   https://ollama.ai
#   ollama pull phi3:mini
#   ollama serve
